Cart Code behind done

+++ ShoppingCartNew/Models/Product.cs
@@ -20,7 +20,7 @@ namespace ShoppingCartNew.Models
         public string ImagePath { get; set; }
 
         [Display(Name = "Price")]
-        public double? UnitPrice { get; set; }
+        public decimal? UnitPrice { get; set; }
 
         public int? CategoryID { get; set; }
